use rsync for copying build, create flags for thoughts script

+++ thoughts.sh
@@ -1,6 +1,3 @@
-trap "exit" INT TERM ERR
-trap "kill 0" EXIT
-
 red="\e[0;91m"
 blue="\e[0;94m"
 green="\e[0;92m"
@@ -8,38 +5,73 @@ reset="\e[0m"
 
 thoughts="${green}@@@ thoughts: "
 
-export AWS_PROFILE=thoughts
+cleanup() {
+    if [ -n "$apiprocess" ]; then kill $apiprocess; fi
+    if [ -n "$interfaceprocess" ]; then kill $interfaceprocess; fi
+}
 
-printf "${thoughts}${blue}Starting wm-api...${reset}"
+trap cleanup EXIT
 
-cd wm-api && npm run start &
-apiprocess=$!
+# build the output and sync with object storage
+build() {
+    export AWS_PROFILE=thoughts
 
-sleep 2
+    printf "${thoughts}${blue}Starting wm-api...${reset}"
+    cd wm-api 
+    npm run start &
+    apiprocess=$!
+    cd ..
+    sleep 2
 
-printf "${thoughts}${blue}Copying images to static directory...${reset}\n"
+    printf "${thoughts}${blue}Copying images to static directory...${reset}\n"
+    mkdir -p ltm-generator/static/images/
+    rsync -rlpgoD --checksum --ignore-existing -v wm/images/ ltm-generator/static/images/
 
-mkdir -p ltm-generator/static/images/
+    printf "${thoughts}${blue}Building static site with ltm-generator...${reset}\n"
+    cd ltm-generator && npm run build &
+    generatorprocess=$!
+    wait $generatorprocess
 
-cp -R wm/images/ ltm-generator/static/images/
+    printf "${thoughts}${blue}Copying build to ltm...${reset}\n"
+    mkdir -p ltm/
+    rsync -rlpgoD --checksum -v --exclude '*.js' --exclude '*.json' ltm-generator/build/ ltm/
 
-printf "${thoughts}${blue}Building static site with ltm-generator...${reset}\n"
+    printf "${thoughts}${blue}Loading config file...${reset}\n"
+    . thoughts.config 
 
-cd ltm-generator && npm run build &
-generatorprocess=$!
+    printf "${thoughts}${blue}Syncing build to object storage...${reset}\n"
+    cd ltm/ && aws s3 sync . $S3_BUCKET --acl public-read --exclude ".DS_Store"
 
-wait $generatorprocess
+    printf "${thoughts}${blue}Export complete, exiting...${reset}\n"
+    exit 
+}
 
-kill $apiprocess
+# run the api and interface 
+write() {
+    printf "${thoughts}${blue}Starting wm-api...${reset}"
+    cd wm-api && npm run start &
+    apiprocess=$!
+    sleep 2
 
-printf "${thoughts}${blue}Loading config file...${reset}\n"
+    printf "${thoughts}${blue}Starting wm-interface...${reset}"
+    cd wm-interface && npm run dev &
+    interfaceprocess=$!
+    sleep 2
 
-. thoughts.config
+    printf "${thoughts}${blue}Ready for writing!${reset}\n"
+    wait $apiprocess
+}
 
-printf "${thoughts}${blue}Syncing build to object storage...${reset}\n"
 
-cd ltm-generator/build/ && aws s3 sync . $S3_BUCKET --acl public-read --exclude "page-preview/*" --include "page-preview/dist/*"
-
-printf "${thoughts}${blue}Export complete, exiting...${reset}\n"
-
-exit 
+while [ "$1" != "" ]; do
+    case $1 in
+    -b | --build)
+        build
+        ;;
+    -w | --write)
+        write
+        ;;
+    *)
+    esac
+    shift 
+done
